// apps/chataigne/ui/src/lib/modules/audio/sound-card/generated/AudioDeviceInspectorState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AudioBackendStatus } from "./AudioBackendStatus";
import type { AudioBufferPolicy } from "./AudioBufferPolicy";
import type { AudioDeviceCatalogEntry } from "./AudioDeviceCatalogEntry";
import type { AudioDeviceDescriptor } from "./AudioDeviceDescriptor";
import type { AudioStreamStatus } from "./AudioStreamStatus";

export type AudioDeviceInspectorState = { discovery_in_progress: boolean, inventory_revision: number, backends: Array<AudioBackendStatus>, device_catalog: Array<AudioDeviceCatalogEntry>, devices: Array<AudioDeviceDescriptor>, input: AudioStreamStatus, output: AudioStreamStatus, engine_sample_rate: number, buffer_policy: AudioBufferPolicy, };